# Add specialization for determining serialized byte sizes.