--- a/Cargo.toml
+++ b/Cargo.toml
@@ -38,6 +38,9 @@ sled = "0.34.7"
 rocksdb = "0.21.0"
 libc = "0.2.99"
 comfy-table = "6.1.0"
+bincode = "1.0"
+serde = "1.0"
+serde_derive = "1.0"
 
 [target.'cfg(target_os = "linux")'.dev-dependencies]
 io-uring = "0.5.1"
@@ -79,4 +82,8 @@ harness = false
 
 [[bench]]
 name = "syscall_benchmark"
+harness = false
+
+[[bench]]
+name = "serde"
 harness = false