Hello,

When reviewing the documentation, this crate is designed for dynamically sized values, however the insert function forces an evaluation of the value to bytes to do a size check.

When using serde, you may have to allocate a Vec if the size is dynamic. This causes an unnecessary Vec to be allocated when using serde to check the size. This patch adds a benchmark for serde as well as specializes the insert function to be able to use whatever sizing function you want and defaults to using the fixed size attribute or the old behavior.

Before:
Bulk loaded 10000 Dynamic elements (50075000 bytes) in 379ms

After:
Bulk loaded 10000 Dynamic elements (50075000 bytes) in 312ms

Speedup: ~17%


## Other benchmarks:

Seems to not affect the benchmarks in general because the default implementation is roughly the same as what was there before.

### Bulk Load

Before:

```
+-----------+--------+-------+---------+--------+-----------+
|           | redb   | lmdb  | rocksdb | sled   | sanakirja |
+===========================================================+
| bulk load | 1725ms | 541ms | 5587ms  | 4766ms | 557ms     |
| bulk load (2MB values) | 14278ms | 8639ms | 25223ms | 42551ms |
| bulk load                 | 2928ms | 1280ms | 7444ms  | 7576ms | 1330ms    |
|---------------------------+--------+--------+---------+--------+-----------|
| individual writes         | 25ms   | 38ms   | 37ms    | 27ms   | 41ms      |
|---------------------------+--------+--------+---------+--------+-----------|
| batch writes              | 1718ms | 879ms  | 350ms   | 923ms  | 1396ms    |
|---------------------------+--------+--------+---------+--------+-----------|
| random reads              | 1243ms | 730ms  | 4487ms  | 2230ms | 975ms     |
|---------------------------+--------+--------+---------+--------+-----------|
| random reads              | 1136ms | 727ms  | 4528ms  | 2021ms | 978ms     |
|---------------------------+--------+--------+---------+--------+-----------|
| random range reads        | 3080ms | 1283ms | 7980ms  | 6389ms | 1492ms    |
|---------------------------+--------+--------+---------+--------+-----------|
| random range reads        | 3058ms | 1281ms | 8131ms  | 6287ms | 1493ms    |
|---------------------------+--------+--------+---------+--------+-----------|
| random reads (4 threads)  | 512ms  | 213ms  | 1821ms  | 735ms  | 572ms     |
|---------------------------+--------+--------+---------+--------+-----------|
| random reads (8 threads)  | 337ms  | 122ms  | 1088ms  | 404ms  | 593ms     |
|---------------------------+--------+--------+---------+--------+-----------|
| random reads (16 threads) | 320ms  | 70ms   | 695ms   | 245ms  | 794ms     |
|---------------------------+--------+--------+---------+--------+-----------|
| random reads (32 threads) | 195ms  | 41ms   | 524ms   | 153ms  | 11218ms   |
|---------------------------+--------+--------+---------+--------+-----------|
| removals                  | 2216ms | 873ms  | 3516ms  | 2916ms | 1370ms    |
+---------------------------+--------+--------+---------+--------+-----------+
```

After:
```
+-----------+--------+-------+---------+--------+-----------+
|           | redb   | lmdb  | rocksdb | sled   | sanakirja |
+===========================================================+
| bulk load | 1769ms | 539ms | 5478ms  | 4510ms | 547ms     |
| bulk load (2MB values) | 13654ms | 8458ms | 24619ms | 42652ms |
| bulk load                 | 2940ms | 1268ms | 7368ms  | 7622ms | 1350ms    |
|---------------------------+--------+--------+---------+--------+-----------|
| individual writes         | 25ms   | 39ms   | 37ms    | 28ms   | 39ms      |
|---------------------------+--------+--------+---------+--------+-----------|
| batch writes              | 1725ms | 863ms  | 357ms   | 917ms  | 1368ms    |
|---------------------------+--------+--------+---------+--------+-----------|
| random reads              | 1219ms | 712ms  | 4682ms  | 2223ms | 979ms     |
|---------------------------+--------+--------+---------+--------+-----------|
| random reads              | 1112ms | 712ms  | 4706ms  | 2035ms | 978ms     |
|---------------------------+--------+--------+---------+--------+-----------|
| random range reads        | 3023ms | 1247ms | 8079ms  | 6378ms | 1503ms    |
|---------------------------+--------+--------+---------+--------+-----------|
| random range reads        | 3024ms | 1247ms | 8038ms  | 6274ms | 1503ms    |
|---------------------------+--------+--------+---------+--------+-----------|
| random reads (4 threads)  | 476ms  | 209ms  | 1900ms  | 684ms  | 573ms     |
|---------------------------+--------+--------+---------+--------+-----------|
| random reads (8 threads)  | 313ms  | 118ms  | 1094ms  | 399ms  | 529ms     |
|---------------------------+--------+--------+---------+--------+-----------|
| random reads (16 threads) | 315ms  | 69ms   | 688ms   | 251ms  | 688ms     |
|---------------------------+--------+--------+---------+--------+-----------|
| random reads (32 threads) | 194ms  | 41ms   | 512ms   | 154ms  | 11152ms   |
|---------------------------+--------+--------+---------+--------+-----------|
| removals                  | 2197ms | 865ms  | 3472ms  | 2986ms | 1387ms    |
+---------------------------+--------+--------+---------+--------+-----------+
```